add some documents for code / thr interests-list part of movie is finished.

+++ Douban.UWP.NET/Tools/Converters/ValueConverter.cs
@@ -8,14 +8,14 @@ using Windows.UI.Xaml.Data;
 namespace Douban.UWP.Tools.Converters {
     public class ValueConverter : IValueConverter {
         public object Convert(object value, Type targetType, object parameter, string language) {
-            return ToValueCode(System.Convert.ToDouble(value), System.Convert.ToUInt32(parameter));
+            return ToValueCode(System.Convert.ToDouble(value), System.Convert.ToDouble(parameter));
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, string language) {
             throw new NotImplementedException();
         }
 
-        private double ToValueCode(double num, uint seed) {
+        private double ToValueCode(double num, double seed) {
             return num / seed;
         }
     }
